fix: stabilize reseller credential rotation SQL

+++ backend/internal/reseller/catalog_refresh.go
@@ -46,7 +46,7 @@ func (s *Service) refreshProductSnapshots(ctx context.Context, resellerID int64)
 		 platform=$3,cost_rate_multiplier=$4,model_snapshot=$5::jsonb,
 		 capabilities=jsonb_set(COALESCE(rp.capabilities,'{}'::jsonb),'{platform}',to_jsonb($3::text)),
 		 price_catalog_version=rp.price_catalog_version+1,effective_at=NOW(),updated_at=NOW()
-		 FROM groups g WHERE rp.id=$1 AND rp.moshu_group_id=$2 AND g.id=$2
+			 FROM groups g WHERE rp.id=$1::bigint AND rp.moshu_group_id=$2::bigint AND g.id=$2::bigint
 		 AND g.deleted_at IS NULL AND g.subscription_type='standard'
 		 AND g.platform=$3 AND g.rate_multiplier=$4 AND COALESCE(g.model_allowlist,'{}'::jsonb)=$6::jsonb
 		 AND (rp.platform IS DISTINCT FROM $3 OR rp.cost_rate_multiplier IS DISTINCT FROM $4 OR rp.model_snapshot IS DISTINCT FROM $5::jsonb)`, item.id, item.groupID, item.platform, item.rate, models, item.raw)
